Ensure rapuma can run from anywhere
When rapuma is started as "scripts/rapuma" from the top folder of the
Git repo, the logic for finding rapumaBase fails, because it's expecting
two directory parts in sys.argv[0] and only finding one. Changing the
logic to instead use os.path.join(dir, '..') instead of the second
dirname() call will work even in the "scripts/rapuma" case.