Fix bug where copying styles would fail
Without this, running:

./scripts/rapuma group FOO mb group add --component_type usfm --cid_list jas --source_list resource/example/MBJAS/sources/mb/jas.usfm

Would fail with the following error:
IOError: [Errno 2] No such file or directory: '/home/rmunn/Publishing/FOO/Style/usfmTex.sty'

What is happening is that it's trying to use shutil.copy() to a
directory that has not yet been created. We could use os.makedirs() for
this, but then we'd have to catch the errno.EEXIST error every time when
we tried to make the directory when it *did* exist. Better to just add
our own makedirs() function in tools.py and use that.