Fix bug with creating projects
The project config was being created on first read (as per the logic in
loadConfig in tools.py, which merges the config with the default config
and writes it out again if there are any changes -- and on first project
creation, the config is empty so of *course* merging with the default
config produces changes), and the test to see if the project already
existed was just testing for an empty directory. Which failed because a
barebones project config (containing only default values) was already in
the project directory. So instead, we need to check if the project
directory has anything *other* than the project config in it. If the
only file in the directory is the project config, this is an empty
project (in the process of being created) and so the creation process
should not say "project already exists, halting". Only if there are
files besides the project config in the directory should the project
creation process think that the project already exists.