Typo fix: resouce -> resource
Since this occurs in config files as well, we need to check config files
for the old setting and convert it to the correct spelling. At some
point in the future when we're confident that all config files with the
old spelling have been converted, that code may be able to be removed.